requirements.txt
langchain
langchain_community
#openai
huggingface_hub
python-dotenv
streamlit